ad_network/requirements.txt
pandas
pyarrow
requests
google-cloud-bigquery
google-cloud-storage
